+++ src/main/java/pw/testoprog/bookingo/repositories/ServiceTypeRepository.java
@@ -3,6 +3,6 @@ package pw.testoprog.bookingo.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pw.testoprog.bookingo.models.ServiceType;
 
-public interface ServiceTypeRepository extends JpaRepository<ServiceType, Long> {
+public interface ServiceTypeRepository extends JpaRepository<ServiceType, Integer> {
 
 }
